+++ src/main.rs
@@ -4,6 +4,7 @@ pub mod codegen;
 pub mod lexer;
 pub mod limits;
 pub mod parser;
+pub mod semantic;
 pub mod string_storage;
 
 use clap::Parser;